feat: organice un bug en el dashboard y le añadi el menu en el encabezado de usuario, cita y categoria
